requirements.txt
openai==1.3.0
python-dotenv>=1.0.0
transformers==4.36.0
anthropic==0.3.11
colorama==0.4.6
prompt-toolkit==3.0.38
rich==13.6.0
pydantic==2.4.2
google-generativeai>=0.3.0
cohere>=4.32
requests>=2.31.0
flask>=2.3.0
flask-wtf>=1.2.0 